split out asm statements parsign
